Pre: Move ByteBuffer.kt to a new shared library for kotlin code
