+++ components/fxa-client/android/build.gradle
@@ -105,6 +105,7 @@ dependencies {
     implementation 'net.java.dev.jna:jna:4.5.2@aar'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'com.google.protobuf:protobuf-lite:3.0.0'
+    implementation project(':as-support-library')
 }
 
 afterEvaluate {
