+++ components/support/android/build.gradle
@@ -1,0 +1,43 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
+
+android {
+    compileSdkVersion 27
+
+    defaultConfig {
+        minSdkVersion rootProject.ext.build['minSdkVersion']
+        targetSdkVersion rootProject.ext.build['targetSdkVersion']
+
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            consumerProguardFiles "$rootDir/proguard-rules-consumer-jna.pro"
+        }
+
+        withoutLib {
+            initWith release
+        }
+    }
+
+}
+
+dependencies {
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+    api 'net.java.dev.jna:jna:4.5.2@aar'
+    api 'com.google.protobuf:protobuf-lite:3.0.0'
+}
+
+apply from: "$rootDir/publish.gradle"
+ext.configurePublish(
+        'org.mozilla.appservices',
+        'support',
+        'Shared types for app services kotlin',
+        /* jnaForTestConfiguration= */ null,
+        /* variantWithoutLib= */ 'androidWithoutLib',
+)
