+++ settings.gradle
@@ -5,11 +5,13 @@ include ':fxa-client-library'
 include ':logins-library'
 include ':places-library'
 include ':rustlog-library'
+include ':as-support-library'
 
 project(':fxa-client-library').projectDir = new File("components/fxa-client/android")
 project(':logins-library').projectDir = new File("components/logins/android")
 project(':places-library').projectDir = new File("components/places/android")
 project(':rustlog-library').projectDir = new File("components/rc_log/android")
+project(':as-support-library').projectDir = new File("components/support/android")
 
 include ':lockbox-megazord'
 project(':lockbox-megazord').projectDir = new File("megazords/lockbox/android")
